changed ledgacy - long instr_ptr type to size_t type, this was causing an issue on win arm64

+++ core/vm/arch/jit/arm64/jit_arm_a64.cpp
@@ -4562,7 +4562,7 @@ void JitRuntime::Initialize(StackProgram* p)
 }
 
 // Executes machine code
-long JitRuntime::Execute(StackMethod* method, size_t* inst, size_t* op_stack, long* stack_pos,
+long JitRuntime::Execute(StackMethod* method, size_t* inst, size_t* op_stack, size_t* stack_pos,
                           StackFrame** call_stack, long* call_stack_pos, StackFrame* frame)
 {
   const int32_t cls_id = method->GetClass()->GetId();
